Rename class to Person to better abstract Leader. To add tests
