+++ src/main/java/seedu/address/model/person/TempPhone.java
@@ -1,4 +1,4 @@
-package seedu.address.model.teammate;
+package seedu.address.model.person;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
