+++ src/main/java/seedu/address/model/person/exceptions/TeammateNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.teammate.exceptions;
+package seedu.address.model.person.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified project.
